fix: confine /search path to data dir too; share the guard
Apply the same data-directory confinement to the /search handler, whose
client-supplied `imagePath` was read by the embedder. Extract the
resolveWithinDataDir guard into utils/util.ts so deleteImage and query
share one implementation, and cover the search traversal-rejection path.
