+++ server/query.ts
@@ -1,7 +1,7 @@
 /* eslint-disable import/no-extraneous-dependencies */
 import { Pinecone, type Index } from '@pinecone-database/pinecone';
 import { embedder } from "./embeddings.ts";
-import { getEnv } from "./utils/util.ts";
+import { getEnv, resolveWithinDataDir } from "./utils/util.ts";
 
 export type Metadata = {
   imagePath: string;
@@ -17,7 +17,9 @@ const getIndex = (): Index<Metadata> => {
 };
 
 const queryImages = async (imagePath: string) => {
-  const queryEmbedding = await embedder.embed(imagePath);
+  // Confine the client-supplied path to the data directory before reading it.
+  const safePath = resolveWithinDataDir(imagePath);
+  const queryEmbedding = await embedder.embed(safePath);
   const queryResult = await getIndex().namespace('default').query({
       vector: queryEmbedding.values,
       includeMetadata: true,
